Enable GitHub Actions for testing LuaVela on Linux
This commit introduces GitHub Actions workflow to test LuaVela on Linux
(Ubuntu 20.04 in particular) against different build types (i.e. Debug
and Release), popular C compilers (i.e. GCC and Clang), and libujit type
(i.e. static or shared) to be used in LuaVela binary built for tests.

Ubuntu 20.04 is chosen due to the verified list of the dependencies.
Anyway, Ubuntu 22.04 and macOS strategy matrices will be introduced
later, when the similar list of the required dependencies is created.

Signed-off-by: Igor Munkin <imun@cpan.org>