Fix key duplication in lj_tab_toset
Before the patch <lj_tab_toset> uses <lj_tab_newkey> helper to insert
the new key into resulting set. As a result, the new slot in GCtab
object is allocated unconditionally (i.e. without the check whether the
given key is already stored in this table) causing key duplication in
the resulting table violating "set" definition.

Within the patch <lj_tab_newkey> is replaced with <lj_tab_set> helper,
that yields the existing slot if the given key is stored in the table,
or allocate the new slot otherwise.

Fixes #38

Signed-off-by: Igor Munkin <imun@cpan.org>