Fix profile ID buffer overflow
The auxiliary buffer for conversion profile ID to string silently
overflows via <sprintf> function. Here is some math below:
* 2 chars are reserved for hexadecimal prefix (i.e. "0x" due to "x"
  format specifier)
* 18 chars are reserved for profile ID (consider preceding zero padding)
* 1 char is required for NUL terminator of the resulting string
  according to <sprintf> manual page

However, the maximum length of 64-bit hexadecimal number is 16, so there
is no need in so many leading zeros in the format. Let's reduce the
padding by one zero to free the slot for the terminator.

Signed-off-by: Igor Munkin <imun@cpan.org>