Simulate system clock for iprof tests
After enabling GitHub Actions to run all uJIT tests by the commit
cd557a9 ("Enable GitHub Actions for testing LuaVela on Linux"), iprof
tests fails from time to time. As a result of the investigation it was
found that only the tests using the system tools to obtain and manage
the time are affected. Since there is no way to tweak CI public runners
provided by GitHub, we can't rely on the environment and have to mock
time-related interfaces used in the aforementioned tests.

To reduce the dependency on the system clock, two libc functions are
overridden in scope of this patch:
* <sleep(3)> -- simply increments second counter stored in the clock
  mock by the given amount of seconds.
* <clock_gettime(3)> -- increments nanosecond counter stored in the
  clock mock by the JIFFY value (i.e. 1 ns, for more info see the
  comment right above the constant definition), normalizes the resulting
  counter value, incrementing second counter if necessary, and fills the
  given struct timespec argument with the "current" time.

Considering everything written above, EPSILON value used as a margin for
float comparison is also reduced to 1e-6 (there is no need in such
relaxed time comparison anymore).

Fixes #41

Signed-off-by: Igor Munkin <imun@cpan.org>