Fix bugprone-suspicious-include warnings
Since Clang Tidy 11.0.0 <bugprone-suspicious-include>[1] check is
introduced. The check detects various cases when an include refers to
what appears to be an implementation file, which often leads to
hard-to-track-down one definition rule violations.

Including source files is widely used in tests to check the particular
API without linking uJIT library, so this check is suppressed except the
two tests below:
* tests/impl/uJIT-tests-C/suite/test_str.c: no need in utils/lj_char.c;
* tests/impl/uJIT-tests-C/suite/test_cci_immutable.c: jit/lj_ir.c is
  replaced with jit/lj_ircall.h;

[1]: https://releases.llvm.org/11.0.0/tools/clang/tools/extra/docs/clang-tidy/checks/bugprone-suspicious-include.html

Signed-off-by: Igor Munkin <imun@cpan.org>