Remove unused variable
Clang Tidy detected the variable with the dead store to it. The variable
is unused since very first uJIT commit, when all 32-bit related code,
introduced in the LuaJIT upstream in scope of the commit e6093b129 ("Fix
register allocation for 8 bit stores in x86 backend."), was removed, but
*these* two occurrences, that were not under !LJ_64 condition, has been
left untouched. So, this is nothing more than a typo.

Signed-off-by: Igor Munkin <imun@cpan.org>