Suppress bugprone-not-null-terminated-result checks
The check[1] is introduced in Clang Tidy 10.0.0 to find function calls
where it is possible to cause a not null-terminated result. It's worth
to mention the fact, that the payload of the Lua string object is not
obliged to be a null-terminated chunk, so the check might spot some
false-positive issues. Anyway, the both found issues worth to be
re-checked whether this is a real issue or just a noise by sanitizer.

As a result, <bugprone-not-null-terminated-result> check is disabled for
both tools/parse_profile/ujpp_demangle_lua.c and src/dump/uj_dump_bc.c
and will be investigated in scope of the #52.

[1]: https://releases.llvm.org/11.0.0/tools/clang/tools/extra/docs/clang-tidy/checks/bugprone-not-null-terminated-result.html

Signed-off-by: Igor Munkin <imun@cpan.org>