Suppress bugprone-branch-clone checks
The check[1] is introduced in Clang Tidy 9.0.0 to find the repeated
branches in if/else-if/else chains, consecutive repeated branches in
switch statements and identical true and false branches in conditional
operators.

The risk of breaking things looks to be much higher than a potential
benefit from fixing away these warnings, so <bugprone-branch-clone>
check is disabled globally, since there is lots of warnings produced by
the sanitizer for translation units as well as for the headers, and will
be investigated in scope of the #53.

[1]: https://releases.llvm.org/11.0.0/tools/clang/tools/extra/docs/clang-tidy/checks/bugprone-branch-clone.html

Signed-off-by: Igor Munkin <imun@cpan.org>