Suppress bugprone-narrowing-conversions checks
The check[1] is introduced in Clang Tidy 7.0.0 to find the silent
narrowing conversions. Though the aforementioned check is an alias for
<cppcoreguidelines-narrowing-conversions> check (so the link to the docs
follows to this warning instead of the original one), it might report
some spots worth to be fixed.

The risk of breaking things looks to be much higher than a potential
benefit from fixing the warnings, so <bugprone-narrowing-conversions>
check is disabled globally, since there is lots of warnings produced by
the sanitizer for translation units as well as for the headers, and will
be investigated in scope of the #54.

[1]: https://releases.llvm.org/11.0.0/tools/clang/tools/extra/docs/clang-tidy/checks/cppcoreguidelines-narrowing-conversions.html

Signed-off-by: Igor Munkin <imun@cpan.org>