Suppress bugprone-signed-char-misuse checks
The check[1] is introduced in Clang Tidy 10.0.0 to find those
signed-char-to-integer conversions which might indicate a programming
error.

The risk of breaking things looks to be much higher than a potential
benefit from fixing the couple of the reported warnings. Though
<bugprone-signed-char-misuse> check reports little warnings, it is
disabled globally to avoid producing even more runs with custom checks.
All the warnings are filed in the #55 and will be investigated later.

[1]: https://releases.llvm.org/11.0.0/tools/clang/tools/extra/docs/clang-tidy/checks/bugprone-signed-char-misuse.html

Signed-off-by: Igor Munkin <imun@cpan.org>