Suppress misc-no-recursion checks
The check[1] is introduced in Clang Tidy 11.0.0 to find strongly
connected functions (by analyzing the call graph for SCC’s (Strongly
Connected Components) that are loops), diagnoses each function in the
cycle, and displays one example of a possible call graph loop
(recursion).

The risk of breaking things looks to be much higher than a potential
benefit from fixing away these warnings, so <misc-no-recursion> check is
disabled globally, since there is lots of warnings produced by the
sanitizer for translation units as well as for the headers, and will be
investigated in scope of the #56.

[1]: https://releases.llvm.org/11.0.0/tools/clang/tools/extra/docs/clang-tidy/checks/misc-no-recursion.html

Signed-off-by: Igor Munkin <imun@cpan.org>