Suppress clang-analyzer-core.UndefinedBinaryOperatorResult checks
As a result of upgrading to Clang Tidy 11.0.0, the sanitizer yields
another warning related to the aforementioned check[1], that verifies
whether binary operators produce the undefined results.

Though the issue is well localized, the risk of breaking things looks to
be much higher than a potential benefit from fixing away these warnings
now, so <clang-analyzer-core.UndefinedBinaryOperatorResult> check is
disabled only for src/jit/lj_record.c, and will be investigated in scope
of the #58.

[1]: https://releases.llvm.org/11.0.0/tools/clang/tools/extra/docs/clang-tidy/checks/clang-analyzer-core.UndefinedBinaryOperatorResult.html

Signed-off-by: Igor Munkin <imun@cpan.org>