Suppress clang-analyzer-core.NullDereference checks
As a result of upgrading to Clang Tidy 11.0.0, the sanitizer yields
another warning related to the aforementioned check[1], that verifies
whether dereferences of null pointers occur.

Though the issue is well localized, the risk of breaking things looks to
be much higher than a potential benefit from fixing away these warnings
now, so <clang-analyzer-core.NullDereference> check is disabled only for
src/jit/opt/mem.c, src/lib/string.c and src/lj_tab.c, and will be
investigated in scope of the #60.

[1]: https://releases.llvm.org/11.0.0/tools/clang/tools/extra/docs/clang-tidy/checks/clang-analyzer-core.NullDereference.html

Signed-off-by: Igor Munkin <imun@cpan.org>