Change BreakBeforeTernaryOperators option policy
The current chosen formatting policy for BreakBeforeTernaryOperators
(i.e. <false>) is broken since Clang Format 11.0.0, causing a misindent
for the sources following the line break. What's even more upsetting is
that this misindent is fixed only in Clang Format 14.0.0. So, this bug
makes us to reformat all the ternary operators now, and revert this
change after upgrading Clang toolchain to 14.0.0.

I believe, that the chosen policy is not widely used, otherwise it would
not be broken for three major releases. For comparison, if this options
is set to <true>, the code is fine for all Clang Format versions from
6.0.0 to 18.0.0. Furthermore, I still believe the code is more readable
with such Clang Format configuration, so I suggest to switch on it.

Signed-off-by: Igor Munkin <imun@cpan.org>