Suppress -Wgnu-null-pointer-arithmetics for Clang
Clang 15 changes the behaviour of -Wnull-pointer-arithmetics to control
-Wgnu-null-pointer-arithmetics either. Since uJIT is built with -Wextra
(including -Wnull-pointer-arithmetics) and -Werror flags being set in
CMAKE_C_FLAGS list, compilation of src/utils/lj_alloc.c fails due to
NULL pointer arithmetics in TOP_FOOT_SIZE constant definition.
Furthermore, -Wno-gnu flag to silence GNU extension diagnostics for
pointer arithmetic is allowed since Clang 15 (llvm/llvm-project#54444).

Considering everything above, tweak the CMAKE_C_FLAGS list for Clang
toolchain only for Clang 15.0.0 and later.

It's worth to mention, that fixing the root cause of the warning is also
an option to fix the issue, but it's better to be in sync with the LuaJIT
upstream at this point.

Signed-off-by: Igor Munkin <imun@cpan.org>