Introduce macOS toolchain setup for CMake
It was mentioned in the commit 89ac4bf82ca30fce4df53f8f534de09cc013f474
("Upgrade to macOS 12 in GitHub Actions"), macos-13 runners use Clang 15
compiler toolchain, that requires to tweak compiler flags the same way
it was done in the commit 3ac8f71517218ba9733961fe814e7693c13cd6d0
("Suppress -Wgnu-null-pointer-arithmetics for Clang"). At the same time,
we only support the one version on macOS (unlike Linux distributions),
so the compiler to be used is well-defined.

Taking into account all of the above, this patch introduces the new
CMake toolchain file with a tiny tweak (and a huge reasoning for it) to
make AppleClang happy on macOS-13. There are no other issues found by
the upgrade, but the whole macOS pipeline still should be refined.

Follows up #63

Signed-off-by: Igor Munkin <imun@cpan.org>