Upgrade to macOS 13 in GitHub Actions
GitHub announced in actions/runner-images#10721 that macos-12 has been
deprecated and will no longer be available after 03.12.2024, so macos-13
will be used in uJIT CI as a result of the patch.

For the history: the part of the docs, referring macOS 11 runners
deprecation has gone (and now I feel like an id10t, reading the commit
89ac4bf82ca30fce4df53f8f534de09cc013f474 ("Upgrade to macOS 12 in GitHub
Actions"), so I decided to mention the particular issue in this one.
Besides, I have no idea, what Microsuck guys will do next, so I decide
to leave a link to the related blogpost[1] either.

[1]: https://github.blog/changelog/2024-08-19-notice-of-upcoming-deprecations-and-breaking-changes-in-github-actions-runners

Signed-off-by: Igor Munkin <imun@cpan.org>