Introduce bump.sh
Finally decided to share my auxiliary script for bumping copyright year
to the public domain: maybe someone would like to replace me in this
annual maintenance, or even setup a scheduled CI pipeline for the next
year. bump.sh is pretty simple: if one runs the script, all flowerboxes
will be adjusted and the changes will be committed to the current
working branch. At the same time, bump.sh is quite Gentoo-ish: it
provides -p (p is for pretend) flag, so the flowerboxes are adjusted,
but the changes are not committed (only staged). Enjoy!

Signed-off-by: Igor Munkin <imun@cpan.org>