Fix moving dead slots in FFI finalizers table
LuaVela has the internal weak table to keep the mapping for GCcdata
objects to their finalizers, which are set via <ffi.gc>. When GCcdata
object becomes dead, there is still the key in this weak table until
it's finalized by the Lua garbage collector (more precisely, until the
next GCfinalize phase). When the table reallocation occurs, these keys
are copied intact to the new hash part.

LuaVela uses <copyTV> for such move, but there is just a plain pointer
assignment in LuaJIT. As a result, the assertion within <tvchecklive>
(to be more precisely within <isdead> check) hits.

This patch replaces <copyTV> with the simple pointer move.

Relates to tarantool/tarantool#5101

Co-authored-by: Igor Munkin <imun@cpan.org>
Signed-off-by: Igor Munkin <imun@cpan.org>