Drop FATAL_ERROR option for cmake_minimum_required
According to the docs[1], the FATAL_ERROR option is accepted but ignored
by CMake 2.6 and higher. It should be specified so CMake versions 2.4
and lower fail with an error instead of just a warning.

Since compatibility with versions of CMake older than 2.8.12 is already
deprecated, there is no need to provide this option anymore.

[1]: https://cmake.org/cmake/help/latest/command/cmake_minimum_required.html

Signed-off-by: Igor Munkin <imun@cpan.org>