Bump CMake minimal version up to 3.5
According to the docs[1], compatibility with versions of CMake older
than 3.5 is deprecated since the version 3.27. Ubuntu 24.04 provides
CMake 3.28.3 by default, so uJIT configuration starts spoiling the
diagnostics with the following warning:
| CMake Deprecation Warning at CMakeLists.txt:11 (cmake_minimum_required):
|   Compatibility with CMake < 3.5 will be removed from a future version of
|   CMake.
|
|   Update the VERSION argument <min> value or use a ...<max> suffix to tell
|   CMake that the project does not need compatibility with older versions.

Since there is no particular policies specified in the project CMake
machinery, there is no need to stick to the 3.3.2 version. However,
migration to the newest version requires more precise changes, so this
patch only bumps the minimal required version up to 3.5.

[1]: https://cmake.org/cmake/help/latest/command/cmake_minimum_required.html

Signed-off-by: Igor Munkin <imun@cpan.org>