Use IN_LIST operator instead of list(FIND ...)
According to the docs[1], IN_LIST operator was introduced in CMake 3.3,
but was not used in uJIT CMake machinery previously for some reasons.
Anyway, CMake 3.5 is the minimal required version now, so this patch
replaces list(FIND ...) hack with IN_LIST operator.

It's worth to mention, that there has always been a policy[2] to tweak
IN_LIST behaviour, but actually, it produces no warning if it's not set
to NEW, so setting this policy is omitted in this patch.

[1]: https://cmake.org/cmake/help/latest/command/if.html#in-list
[2]: https://cmake.org/cmake/help/v3.3/policy/CMP0057.html

Signed-off-by: Igor Munkin <imun@cpan.org>